choose clang 5.0.0 or 7.0.0 depending what's installed

+++ .travis/install.sh
@@ -2,10 +2,17 @@
 set -x
 
 # setup clang for clang tests using local clang installation
-sudo ln -s /usr/local/clang-7.0.0/bin/clang /usr/bin/clang
-sudo ln -s /usr/local/clang-7.0.0/bin/clang++ /usr/bin/clang++
 
-ls -l /usr/local/clang-7.0.0/bin/clang* /usr/bin/clang*
+
+if [ ! -f /usr/local/clang-5.0.0/bin/clang ]; then
+    echo "No Clang 5.0.0 trying 7.0.0"
+    sudo ln -s /usr/local/clang-7.0.0/bin/clang /usr/bin/clang
+    sudo ln -s /usr/local/clang-7.0.0/bin/clang++ /usr/bin/clang++
+else
+    echo "Clang 5.0.0"
+    sudo ln -s /usr/local/clang-5.0.0/bin/clang /usr/bin/clang
+    sudo ln -s /usr/local/clang-5.0.0/bin/clang++ /usr/bin/clang++
+fi
 
 # dependencies for gdc tests
 sudo apt-get -y install gdc
